Parameter optimization: Muscollo

+++ tropter/tropter/optimalcontrol/OptimalControlProblem.hpp
@@ -109,7 +109,7 @@ void OptimalControlProblem<T>::
 initialize_on_mesh(const Eigen::VectorXd&) const
 {}
 
-template<typename T>
+template<typename T>    
 void OptimalControlProblem<T>::
 calc_differential_algebraic_equations(const DAEInput<T>&, DAEOutput<T>) const
 {}
@@ -204,7 +204,7 @@ set_parameter_guess(OptimalControlIterate& guess,
     if (guess.parameters.size() == 0) {
         guess.parameters.resize(m_parameter_infos.size());
     }
-    else if (guess.parameters.size() != m_parameter_infos.size()) {
+    else if (guess.parameters.size() != (int)m_parameter_infos.size()) {
         TROPTER_THROW("Expected guess.parameters to have %i elements "
             "but it has %i elements.",
             m_parameter_infos.size(), guess.parameters.size());
